Backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-openai
pypdf
faiss-cpu
python-dotenv